requirements.txt
flask
ultralytics==8.2.0
opencv-python-headless
pillow
